bot/requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
PyJWT==2.10.1
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
